backend/requirements.txt
Flask
torch
torchvision
Pillow
boto3
pymysql